Incorporate code review comments
