+++ src/modules/UI/Wallets/action.js
@@ -40,7 +40,6 @@ export const UPDATE_EXISTING_TOKEN_SUCCESS = 'UPDATE_EXISTING_TOKEN_SUCCESS'
 export const OVERWRITE_THEN_DELETE_TOKEN_SUCCESS = 'OVERWRITE_THEN_DELETE_TOKEN_SUCCESS'
 export const ADD_NEW_TOKEN_THEN_DELETE_OLD_SUCCESS = 'ADD_NEW_TOKEN_THEN_DELETE_OLD_SUCCESS'
 export const UPDATE_WALLET_LOADING_PROGRESS = 'UPDATE_WALLET_LOADING_PROGRESS'
-export const INSERT_WALLET_IDS_FOR_PROGRESS = 'INSERT_WALLET_IDS_FOR_PROGRESS'
 
 export const refreshReceiveAddressRequest = (walletId: string) => (dispatch: Dispatch, getState: GetState) => {
   const state = getState()
@@ -446,19 +445,6 @@ export function updateWalletLoadingProgress (walletId: string, addressLoadingPro
   return {
     type: UPDATE_WALLET_LOADING_PROGRESS,
     data
-  }
-}
-
-export function insertWalletIdsForProgress (activeWalletIds: Array<string>) {
-  const activeWalletIdList = activeWalletIds
-  const activeWalletIdProgress = {}
-  activeWalletIdList.map((item) => {
-    activeWalletIdProgress[item] = 0
-  })
-
-  return {
-    type: INSERT_WALLET_IDS_FOR_PROGRESS,
-    data: activeWalletIdProgress
   }
 }
