+++ src/modules/UI/components/SimpleDropdown/SimpleDropdown.ui.js
@@ -1,15 +1,15 @@
 // @flow
 
 import React, { Component } from 'react'
-import { TouchableWithoutFeedback } from 'react-native'
+import { TouchableWithoutFeedback, StyleSheet } from 'react-native'
 import styles from './style.js'
 import * as Animatable from 'react-native-animatable'
 
 export type Props = {
   height?: number,
   children: any,
-  onPress: Function,
-  containerStyle: Object
+  onPress: () => void,
+  containerStyle: StyleSheet.Styles
 }
 
 export type State = {
@@ -19,7 +19,7 @@ export type State = {
 export default class WalletListProgressDropdown extends Component<Props> {
   render () {
     return (
-      <Animatable.View animation={'fadeIn'} duration={6000} /* slideInDown */ style={[styles.container, this.props.containerStyle, {height: this.props.height}]}>
+      <Animatable.View animation={'fadeIn'} duration={6000} style={[styles.container, this.props.containerStyle, {height: this.props.height}]}>
         <TouchableWithoutFeedback onPress={this.props.onPress}>
           {this.props.children}
         </TouchableWithoutFeedback>
